コントローラーレシーバー追加
