+++ Packages/com.nomlas.topazchat/Runtime/Scripts/ControllerReceiver.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e101b2b8602ce6f4fade64ac3edf567d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
