+++ Packages/com.nomlas.topazchat/Runtime/Scripts/URLSync.cs
@@ -4,7 +4,7 @@ using VRC.SDKBase;
 
 namespace Nomlas.TopazChat
 {
-    public class URLSync : VideoEventListener
+    public class URLSync : ControllerReceiver
     {
         // ----------------------------------------
